Create the `User` model
Getting the database stuff wired in is always a bit of a challenge; so I
tend to make the first model and get that working before making any of
the UI changes.

I tested this out by starting irb and running:

```
require_relative 'models'
user = User.create({:name => "Zee", :email => "zee@example.com",
:password => "password"})

puts User.first
```

Which proved that my model was wired up and ready to go!

The observant among you may have noticed that my `password` is stored in
plaintext. This is a common anti-pattern:
https://stormpath.com/blog/why-you-might-want-to-store-your-passwords-in-plain-text/

We'll fix that in the next commit