Secure the Users Password Field!
As we've already seen, storing your passwords in plain text isn't a good
idea:
https://stormpath.com/blog/why-you-might-want-to-store-your-passwords-in-plain-text/

To prevent that, we've updated the User model so a password is "hashed"
when it is set. "Hashing" is the act of reproducibly, non-recoverably
encrypting data.

"Woah Woah Woah," you say "Reproducibly, non-recoverably? What madness
is this?" We want passwords to be non-recoverable so if some nefarious
person gains access to your database they can't undo our encryption.

We want passwords to be reproducibly encrypted, so when our users enter
their password, we can verify that it matches the one we've stored.

In this case, the word `bcrypt` is actually two things:

1. an algorithm developed for securely hashing data
1. a ruby gem:
https://github.com/codahale/bcrypt-ruby#how-to-use-bcrypt-ruby-in-general

Play around with bcrypt in `irb`:
```
require 'bcrypt'

password = BCrypt::Password.create("password")
puts "password"
puts password
```
These don't look the same...

`puts password == "password"`
Then why is that's true?!

`puts password.class`
Oh, the `password` variable is not *actually* a String!

`puts password.methods`
Ah, it seems that there's a method on BCrypt::Password called `==`

`puts password.==("password")`
When we use `==` in ruby, we're calling the method `==` on the
variable on the left with whatever is on the right.

`puts "password" == password`
So... Why is this false?!

A variable of the class `BCrypt::Password` can compare a string to
itself; but a string can't compare itself to a `BCrypt::Password`.

COOMPPPUTTEERRRS :'(.

So... what about `BCrypt::Password.new`

When we store a password in the database, it gets saved as a `String`.
When we pull the data back out of the database, we need a way to make it
comparable to non-hashed strings again. Go back to your irb session.

```
password_as_string = password.to_s
puts password_as_string.class
```
Hooray! It's a regular boring old string! That means it shouldn't be
equal to the actual string "password"

`puts password_as_string == "password"`
Whooo! It's false!

```
comparable_password = BCrypt::Password.new(password_as_string)
puts comparable_password == "password"
```

So, there we go. BCrypt lets us store passwords securely in a database;
and then pull them out and compare them without ever giving anyone the
opportunity to figure out what the password actually is.