Handle Password Errors more gracefully
NOTE: Yes I know you can't create a user right now. I wanted to
demonstrate handling errors gracefully!

Here we see a couple things for the first time:

1. Calling the `errors` method on a Model
2. Using a "partial" with a hash of "local" variables
3. Re-using a form and pre-populating it with the data a user entered
   already

On Model Errors
===============
In ruby-land, one of a `Models` jobs is to ensure that bad data cannot
be put into a database. In DataMapper, this is enforced by `validating`
the data before saving a model. When validations fail, an error is added
to the array that lives in the `errors` hash at the key named for the
property that failed validation.

For example, pretend the User model had a validation that required its
name property be at least two words:

```
zee = User.create({ :name => "Zee", :email => "zee@example.com", :password =>
"password" })
puts zee.saved? # => false
puts zee.errors[:name] # => ["Your name must be two (or more!) words!"]
puts zee.errors[:email] # => []
```
Isn't that handy?! We can retrieve a list of errors by the name of the
property!

On Partials
===========
Think of an erb partial almost like a ruby method. It's a way to define
a re-usable template that takes arguments and returns a string.

When we call a partial, we give it two arguments:

1. The name of the view to render
2. A hash of options.

In this case, one of the keys in the hash of options is `:locals`, which
we use to pass named data into the template. If we translated the
partial and the calling of the partial into ruby, it would look
something like this:

```
def error(model, field)
  "do some work here"
end

error(@user, :password)
```

On Re-Rendering a View With the Entered Data
============================================
This is a good habit to get into. You don't want to force your users to
retype their name and email address just because the password
confirmation didn't match!

This is why we have the `@user = User.new` in the `get "/users/new"`
route handler. It ensures we have a @user instance variable to pull
data from (even though there's nothing inside)