Add Validatons for Name and Email Address
Now all that work we did last commit is paying off! All we need to do to
get errors displaying for the name and email fields was:

1. Copy-paste the call to the `_error` partial
2. Change :field_name in the :locals hash to match the field. Done!

The second bit of this commit was adding validations for the User model.
I'm using what DataMapper calls "Auto-validations" by adding some keys
and values to the options hash when I define the property.

Docs: http://datamapper.org/docs/validations.html

The docs above, while a bit dense, exemplify the kind of docs you'll be
expected to read and understand as a professional developer. They're
clear, they have examples, and they're close to exhaustive.

Give the whole thing a skim; don't worry if you don't grok everything.
85% of your job is reading docs and finding the information you need
*right then* without being caught up in the minutia.

Test it Out
===========
Go ahead, try to create two people with the same email address or
without a name! NOPE! CAN't! HA! Take that!