Make form fields proceed down the page, instead of across it.
Look, I'm not a designer. This is literally as pretty as I make things
if left to my own devices.

The important thing to notice is the new top-level directory "public"

The public directory is ignored by sinatra, and served up as-is. This is
where you put images, css, javascript, and other static assets that your
application uses to look beautiful and be all fancy.

You'll notice I've added a `link` tag to the `head` of
`views/layout.erb`. This is how we add stylesheets to a web page.

On Design, HTML, and CSS
========================

For those of you with a hankering to learn design better as well as html
and css:

1. learn.shayhowe.com - seriously. This is your HTML/CSS bible. Know it.
   Love it.
5. Don't Make Me Think - The definitive introduction to application design
3. The non-designers design book - A wonderful introduction to visual design
4. Rocket Surgery Made Easy - An accessible  beginners guide to user
   experience
2. hackdesign.com - A great, easy to follow course on design