Allow Users to log in
Well well well, we're finally getting somewhere interesting!

We've done two things:

1. We've written a log-in method
2. We've started to use sessions to persist data across a single users
   many requests.

On Sessions
===========

Here we've got the notion of "sessions". A session uniquely identify a
given visitor of your website. It lets you persist data over many HTTP
requests. Ever wonder how Google or Facebook or Amazon knows you're you?

Well, they use a session.

We're using a session to store the ID of the user who logged
in; and when we see that we have that information we don't ask people to
sign up or log in on the homepage. Later we'll use it for more fancy
stuff.

The cool thing about sessions is they're *very difficult* to tamper
with. All the data in the session is transmitted to the server and back
to the client on every HTTP request; but it's encrypted!

On Login Procedures
===================
The other bit to keep in mind is the procedure we're using for verifying
log-ins. Because we're storing the password in the database as a string,
we can't just query the email and password at the same time. Instead we
have to retrieve based on the email address, then compare the password
stored in the database to the password the user typed in.

If all that goes as planned, we stick their user id in the session! BAM!
Now we know who that person is even if they close the web page. They
would have to clear their cookies to stop us from knowing who they are.