Registering logs you in
Now we introduce the concept of sinatra helpers. Sinatra Helpers allow
you to define methods that:

1. Have access to `instance variables` defined in your routes
2. May be called in your views

Here I'm using helpers to wrap access to the `session` hash. This allows
me to:

* Load the current_user from the database and cache it once per request;
  even if I access to it several different times.
* `login` a user, I.e. set the session variable and redirect.
* Provide a `logged_in?` helper to reveal our intent better instead of
  asking `if session[:current_user]`

By doing this, I can encourage restricting access to the `session` hash;
which frequently becomes a den of complexity in web applications.