Add a base line of security for sessions
You may have seen the following warning in your server logs:

```
SECURITY WARNING: No secret option provided to Rack::Session::Cookie.
This poses a security threat. It is strongly recommended that you
provide a secret to prevent exploits that may be possible from crafted
cookies. This will not be supported in future versions of Rack, and
future versions will even invalidate your existing user cookies.
```

In order for sessions to be truly secure, we need to use a unique,
random string to encrypt the sessions. This is known as the Session's
"Secret." When we use a secret we increase the security of our
application significantly.