User may log out
This is the final piece in our first story: "Guest may register as user"

The main thing to notice is I'm using a `delete` route and a form to log
a user out; then when that route is called I simply remove the
`:current_user` from the session and *wala!* logged out!

The second thing to notice is I updated my list of requirements. This is
important! When you think you're done with something you shoul update
whatever you're using to track your work. If you're not using anything
to track your work then you should. I like to just use my README for
small projects; but if I'm working on a team I prefer github issues.

I also added a touch of style because, as you know, I'm the greatest web
designer. evar. EVAR!!!11oneone