User may add a restaurant
Most of this was simply duplicating the work we did with the user
registration, with a few minor details around:

1. We expose all the Restaurants on the home page
2. We ensure a Restaurant "belongs to" the User who created it

This is the notion of "Associations." Association are how you declare
relationships between models in DataMapper.

There are three main kinds of associations:

1. One to Many (i.e. One Person has many Tweets)
2. One to One (I.e. One Person has One Profile)
3. Many to Many (I.e. Many People Favorite Many Tweets)

The substance of almost all web applications is in how they associate
data and expose it to the user. Here we're using a 1 to Many
relationship to connect 1 User to Many Restaurants.

Take a moment to skim the DataMapper documentation on Associations:
http://datamapper.org/docs/associations.html

Keep in mind: They're providing *very in depth* usage examples for some
of the less common database relationships; so don't worry about getting
all of it. Focus on the 1 to many and maybe *maybe* the many to many.