Retrieve and store latitude and longitude w/restaurant
Since most people don't *know* the latitude and longitude of their
favorite restaurants (I have no idea why not. Mine is latitude 37.805899
and longitude -122.274485), we have to look them up.

Thankfully, there's a gem for that! `graticule` from Collective Idea
(one of my old coworkers consuting shops). Graticule lets us hook into
Googles Geo-lookup API which returns a latitude and longitude for a
given address. Cool, huh?

This adds in `rake` which is a tool for running 1 off commands that
augment your application; but aren't the actually application itself.

I'll be using it to update the latitude and longitude of the restaurants
we already have in the can-i-eat-here.herokuapp.com database.